Version 0.3 - I have started to develop the Enemy module by adding a sprite that is present on the map.
Success criteria to be completed in this version:
The enemy will go to the player when spawned in
The map must allow AI enemies to spawn in random areas.

I have faced trouble with moving the sprite and making it follow the player. In version 0.3.1, the enemy will be able to move.