Version 0.4.2 In this version, I have made the enemies die after being shot. I did this by parenting the enemy instances to an entity, which can then be used to make the ray casted from the gun only impact enemy instances, which in this case means that they will be removed from the map.
Success criteria fulfilled in this version:
AI enemies will die when shot

In version 0.4.3, I will work on a counter that counts how many enemies die, which will then be saved in a csv or txt file as a form of persistent storage. I will also work on making the game fullscreen to fulfill the success criteria "The window will default to full screen".